Move Scanners to bisq.wallets.regtest module
We can't depend on Bisq 2's common module to be able to reuse the
wallets library in Bisq 1.
